Class Board
Functions added to class board: water_lines
Merged ajacent_values functions in one function that returns a list with all the valid ajacent values of a cell starting with the top adjacent value and going anti-clock through all the reamining values.

Changed the row and col restritions atirbutes of the board class in to two new atributes called row_info and col_info where we store the row and col retritions, numb of empty cells and numb of pieces.